Update main.py
The game has an API integration of Open Trivia DB. The api_call module sends an API request to Open Trivia DB and fetches the True/False questions in form of json and parses it to a list of dictionaries. The user gets a dynamic experience where they can select how many questions they want and what is the difficulty level, these parameters are passed to the API url to fetch the required data.